reworked components
