+++ data/frontend/src/components/Chat/Message/JoinMessage.tsx
@@ -1,0 +1,8 @@
+import { chatResponseArgs } from '../args.interface';
+
+
+export default function JoinMessage({message}: {message: chatResponseArgs}) {
+	return (
+		<p>=&#62; <a href='blank'>{message.source}</a> joined the room</p>
+	);
+}
