+++ data/frontend/src/components/Chat/Message/LeftMessage.tsx
@@ -1,0 +1,7 @@
+import { chatResponseArgs } from '../args.interface';
+
+export default function LeftMessage({message}: {message: chatResponseArgs}) {
+	return (
+		<p>&#60;= <a href='blank'>{message.source}</a> left the room</p>
+	);
+}
